% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif4.R
\name{dif4}
\alias{dif4}
\title{order 4 differencing of a time series vector}
\usage{
dif4(x)
}
\arguments{
\item{x}{{(n X 1) vector of time series (market returns) with n items each}}
}
\value{
ou2 matrix having five columns, first for x, the next four
columns have diff(x), diff-squared(x), diff-cubed(x) and diff-fourth(x)
}
\description{
This is for momentum traders who focus on growth, acceleration, its gorwth
and further acceleration. The diff function of R seems to do 
recycling of available numbers, not wanted for our purposes.
}
\examples{
x=c(2,8,3,5,1,8,19,22,23)
dif4(x)

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
