% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif4mtx.R
\name{dif4mtx}
\alias{dif4mtx}
\title{order four differencing of a matrix of time series}
\usage{
dif4mtx(mtx)
}
\arguments{
\item{mtx}{{(n X p) matrix of p time series (market returns) with n items each}}
}
\value{
out matrix having 12 rows, (data, D1 to D4
and ranks of D1 to D4
The column names of out are those of input matrix mtx.
}
\description{
This is for momentum traders who focus on growth, acceleration, its growth
and further acceleration. The diff function of R seems to do 
recycling of available numbers, not wanted for our purposes. Hence, this
function is needed in portfolio studies based on time series.
}
\examples{
x=c(2,8,3,5,1,8,19,22,23)
y=c(3,11,2,6,7,9,20,25,21)
dif4mtx(cbind(x,y))

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
