% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeq.R
\name{getSeq}
\alias{getSeq}
\title{Two sequences: starting+ending values from n and blocksize (internal use)}
\usage{
getSeq(n, blksiz)
}
\arguments{
\item{n}{{length of the range}}

\item{blksiz}{{blocksize}}
}
\value{
two vectors sqLO and sqUP
}
\description{
This is an auxiliary function for gmcmtxBlk. It gives sequences of starting
and ending values
}
\examples{

getSeq(n=99, blksiz=10)

}
\seealso{
\code{\link{gmcmtxBlk}}
}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
