% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor.R
\name{minor}
\alias{minor}
\title{Function to do compute the minor of a matrix defined by row r and column c.}
\usage{
minor(x, r, c)
}
\arguments{
\item{x}{{the input matrix}}

\item{r}{{the row number}}

\item{c}{{the column number}}
}
\value{
the appropriate `minor' matrix from the original matrix.
}
\description{
Function to do compute the minor of a matrix defined by row r and column c.
}
\note{
this is needed by the cofactor function
}
\examples{

\dontrun{
 x=matrix(1:20,ncol=4)
minor(x,1,2)}

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}

