% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napair.R
\name{napair}
\alias{napair}
\title{Function to do pairwise deletion of missing rows.}
\usage{
napair(x, y)
}
\arguments{
\item{x}{Vector of x data}

\item{y}{Vector of y data}
}
\value{
\item{newx}{A new vector x after removing pairwise missing data} 
\item{newy}{A new vector y after removing pairwise missing data}
}
\description{
The aim in pair-wise deletions is to retain the largest 
number of available data pairs with all non-missing data.
}
\examples{

\dontrun{
x=sample(1:10);y=sample(1:10);x[2]=NA; y[3]=NA
napair(x,y)}

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
