% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naTriplet.R
\name{naTriplet}
\alias{naTriplet}
\title{Function to do matdched deletion of missing rows from x, y and control variable(s).}
\usage{
naTriplet(x, y, ctrl)
}
\arguments{
\item{x}{Vector of x data}

\item{y}{Vector of y data}

\item{ctrl}{{Data matrix on the control variable(s) beyond causal path issues}}
}
\value{
\item{newx}{A new vector x after removing pairwise missing data} 
\item{newy}{A new vector y after removing pairwise missing data}
}
\description{
The aim in three-way deletions is to retain only the largest 
number of available data triplets with all non-missing data.
}
\examples{

\dontrun{
x=sample(1:10);y=sample(1:10);x[2]=NA; y[3]=NA
napair(x,y)}

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
\seealso{
See \code{\link{napair}}.
}

