% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siPair2Blk.R
\name{siPair2Blk}
\alias{siPair2Blk}
\title{Block Version of silentPair2 for causality scores with control variables}
\usage{
siPair2Blk(mtx, ctrl = 0, dig = 6, blksiz = 10)
}
\arguments{
\item{mtx}{{The data matrix with p columns. Denote x1 as the first column, 
which is fixed and then 
 paired with all other columns, say: x2, x3, .., xp, one by one  
 flipping with x1.The number of columns, p, must be 2 or more}}

\item{ctrl}{{data matrix for designated control variable(s) outside causal paths.
The default ctrl=0 means that there are no control variables used.}}

\item{dig}{{Number of digits for reporting (default \code{dig}=6).}}

\item{blksiz}{{block size, default=10, if chosen blksiz >n, where n=rows in the
matrix, then blksiz=n. That is, no blocking is done}}
}
\value{
With p columns in \code{mtx} argument to this function, x1 can be 
paired with a total of p-1 columns (x2, x3, .., xp). Note
we never flip any of the control variables with x1.  This function
produces i=1,2,..,p-1 numbers representing the summary sign, or `sum' from 
the signs sg1 to sg3 associated with the three criteria:
Cr1, Cr2 and Cr3.  Note that sg1 and sg2 themselves are weighted signs using
the weighted sum of signs from four orders of stochastic dominance.
In general, a positive sign in the i-th location of the `sum' output of this function
means that x1 is the kernel cause while the variable in (i+1)-th column of \code{mtx} is the
`effect' or `response' or `endogenous.' The magnitude represents the strength (unanimity)
of the evidence for a particular sign. Conversely, a negative sign
 in the i-th location of the `sum' output of this function means that
that the first variable listed as the input to this function is the `effect,'
while the variable in (i+1)-th column of \code{mtx} is the exogenous kernel cause.
}
\description{
Block version allows a new bandwidth (chosen by the np package)
while fitting kernel regressions for each block of data. This may
not be appropriate in all situations.  Block size is flexible. 
The function develops a unanimity index regarding the particular
flip (y on xi) or (xi on y) is best. Relevant signs determine the
causal direction and unanimity index among three criteria.
The `2' in the name of the function suggests a second implementation
where exact stochastic dominance, decileVote, and momentVote are used.
It avoids Anderson's trapezoidal approximation.
The summary results for all
three criteria are reported in a vector of numbers 
internally called \code{crall}.
}
\note{
The European Crime data has all three criteria correctly suggesting that
high crime rate kernel causes the deployment of a large number of police officers.
The command \code{attach(EuroCrime); silentPairs(cbind(crim,off))}
returns only one number: 3.175, implying the highest unanimity strength index,
with the positive sign suggesting `crim' in the first column kernel causes
`off' in the second column of the argument \code{mtx} to this function.
}
\examples{


\dontrun{
options(np.messages=FALSE)
colnames(mtcars[2:ncol(mtcars)])
siPair2Blk(mtcars[,1:3],ctrl=mtcars[,4:5]) # mpg paired with others
}

options(np.messages=FALSE)
set.seed(234)
z=runif(10,2,11)# z is independently created
x=sample(1:10)+z/10 #x is somewhat indep and affected by z
y=1+2*x+3*z+rnorm(10)
w=runif(10)
x2=x;x2[4]=NA;y2=y;y2[8]=NA;w2=w;w2[4]=NA
siPair2Blk(mtx=cbind(x2,y2), ctrl=cbind(z,w2))


}
\references{
H. D. Vinod  'Generalized Correlation and Kernel Causality with
   Applications in Development Economics' in Communications in
   Statistics -Simulation and Computation, 2015,
   \doi{10.1080/03610918.2015.1122048}

Vinod, H. D. Causal Paths and Exogeneity Tests 
in {Generalcorr} Package for Air Pollution and Monetary Policy 
(June 6, 2017). Available at SSRN: 
\url{https://www.ssrn.com/abstract=2982128}
}
\seealso{
See  \code{\link{bootPairs}}, \code{\link{silentMtx}}

See  \code{\link{someCPairs}}, \code{\link{compPortfo}}
}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY.
}
\concept{Hausman-Wu exogeneity criteria}
\concept{generalized correlations}
\concept{stochastic dominance}
