% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_matrix.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sort all columns of matrix x by j-th column while carrying along all columns.}
\usage{
sort_matrix(x, j)
}
\arguments{
\item{x}{{is a matrix with several columns}}

\item{j}{{column number with which to sort}}
}
\value{
{A sorted matrix}
}
\description{
This function simply uses sort.list function in R.
}
\examples{

set.seed(30)
x=matrix(sample(1:50),ncol=5)
y=sort_matrix(x,3);y
}

