% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdz_xy.R
\name{stdz_xy}
\alias{stdz_xy}
\title{Standardize x and y vectors to achieve zero mean and unit variance.}
\usage{
stdz_xy(x, y)
}
\arguments{
\item{x}{Vector of data which can have NA's}

\item{y}{Vector of data which can have NA's}
}
\value{
\item{stdx}{standardized values of x}
\item{stdy}{standardized values of y}
}
\description{
Standardize x and y vectors to achieve zero mean and unit variance.
}
\note{
This works even if there are missing x or y values.
}
\examples{

\dontrun{
set.seed(30)
x=sample(20:30)
y=sample(21:31)
stdz_xy(x,y) }

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}

