% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryRank.R
\name{summaryRank}
\alias{summaryRank}
\title{Compute ranks of rows of matrix and summarize them into a choice suggestion.}
\usage{
summaryRank(mtx)
}
\arguments{
\item{mtx}{{matrix to be ranked by row and summarized}}
}
\value{
a matrix appending 4 more rows of ranks and a couple of rows
summarizing ranks and also choice priority number for each stock (asset)
}
\description{
This function allows getting out the choice (of column stock) from four
rows of numbers quantifying the four orders of exact stochastic dominance comparisons.
If the choice row has 1 then that column is to be chosen with the largest
(portfolio) weight. If the original matrix row names exist, same are repeated
for the extra rows representing the ranks.  The row name for sum of ranks is
sumRanks and the ranking of the sumRanks providing the choice is named choice
along the bottom row of the output matrix called "out."
}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
