% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor.R
\name{minor}
\alias{minor}
\title{Function to do compute the minor of a matrix defined by row r and column c.}
\usage{
minor(x, r, c)
}
\arguments{
\item{x}{{The input matrix}}

\item{r}{{The row number}}

\item{c}{{The column number}}
}
\value{
The appropriate `minor' matrix defined from the input matrix.
}
\description{
Function to do compute the minor of a matrix defined by row r and column c.
}
\note{
This function is needed by the cofactor function.
}
\examples{

\dontrun{
 x=matrix(1:20,ncol=4)
minor(x,1,2)}

}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY
}
\concept{minor of a matrix}
