% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silentPairs.R
\name{silentPairs}
\alias{silentPairs}
\title{Function for kernel causality into 3-column matrix admitting control variables}
\usage{
silentPairs(mtx, ctrl = 0, dig = 6, wt = c(1.2, 1.1, 1.05, 1),
  sumwt = 4)
}
\arguments{
\item{mtx}{{The data matrix with p columns. Denote x1 as the first column 
which is fixed and then 
 paired with all other columns, say: x2, x3, .., xp, one by one for the 
 purpose of flipping with x1. p must be 2 or more}}

\item{ctrl}{{data matrix for designated control variable(s) outside causal paths}}

\item{dig}{{Number of digits for reporting (default \code{dig}=6).}}

\item{wt}{{Allows user to choose a vector of four alternative weights for SD1 to SD4.}}

\item{sumwt}{{ Sum of weights can be changed here =4(default).}}
}
\value{
With p columns in \code{mtx} argument to this function, x1 can be 
paired with a total of p-1 columns (x2, x3, .., xp). Note
we never flip any of the control variables with x1.  This function
produces i=1,2,..,p-1 numbers representing the summary sign, or `sum' from 
the signs sg1 to sg3 associated with the three criteria:
Cr1, Cr2 and Cr3.  Note that sg1 and sg2 themselves are weighted signs using
weighted sum of signs from four orders of stochastic dominance.
In general, a positive sign in the i-th location of the `sum' output of this function
means that x1 is the kernel cause while the variable in (i+1)-th column of \code{mtx} is the
`effect' or `response' or `endogenous.' The magnitude represents the strength (unanimity)
of the evidence for a particular sign. Conversely a negative sign
 in the i-th location of the `sum' output of this function means that
that the first variable listed as the input to this function is the `effect,'
while the variable in (i+1)-th column of \code{mtx} is the exogenous kernel cause.
This function is a summary of \code{someCPairs} 
allowing for control variables.
}
\description{
Allowing input matrix of control variables and missing data, this function produces a
 3 column matrix summarizing the results where the estimated signs of
stochastic dominance order values (+1, 0, -1) are weighted by 
\code{wt=c(1.2,1.1, 1.05, 1)} to
compute an overall result for all orders of stochastic dominance by a weighted sum for
the criteria Cr1 and Cr2 and added to the Cr3 estimate as: (+1, 0, -1).
}
\details{
The reason for slightly declining weights on the signs from
SD1 to SD4 is simply that the local mean comparisons 
implicit in SD1 are known to be
more reliable than local variance implicit in SD2, local skewness implicit in
SD3 and local kurtosis implicit in SD4. The source of slightly declining sampling
unreliability of higher moments is the
higher power of the deviations from the mean needed in their computations.
The summary results for all
three criteria are reported in a vector of numbers internally called \code{crall}:
}
\note{
The European Crime data has all three criteria correctly suggesting that
high crime rate kernel causes the deployment of a large number of police officers.
The command \code{attach(EuroCrime); silentPairs(cbind(crim,off))}
returns only one number: 3.175, implying the highest unanimity strength index,
with the positive sign suggesting `crim' in the first column kernel causes
`off' in the second column of the argument \code{mtx} to this function.
}
\examples{


\dontrun{
options(np.messages=FALSE)
colnames(mtcars[2:ncol(mtcars)])
silentPairs(mtcars[,1:3],ctrl=mtcars[,4:5]) # mpg paired with others
}

options(np.messages=FALSE)
set.seed(234)
z=runif(10,2,11)# z is independently created
x=sample(1:10)+z/10 #x is somewhat indep and affected by z
y=1+2*x+3*z+rnorm(10)
w=runif(10)
x2=x;x2[4]=NA;y2=y;y2[8]=NA;w2=w;w2[4]=NA
silentPairs(mtx=cbind(x2,y2), ctrl=cbind(z,w2))


}
\references{
H. D. Vinod 'Generalized Correlation and Kernel Causality with
   Applications in Development Economics' in Communications in
   Statistics -Simulation and Computation, 2015,
   \url{http://dx.doi.org/10.1080/03610918.2015.1122048}
}
\seealso{
See  \code{\link{bootPairs}}.

See  \code{\link{someCPairs}}, \code{\link{some0Pairs}}
}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY.
}
\keyword{SD1,}
\keyword{SD2,}
\keyword{SD3,}
\keyword{SD4,}
\keyword{causal}
\keyword{correlations}
\keyword{criteria,}
\keyword{generalized}
