% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_matrix.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sort all columns of matrix x with respect to the j-th column.}
\usage{
sort_matrix(x, j)
}
\arguments{
\item{x}{{An input matrix with several columns}}

\item{j}{{The column number with reference to which one wants to sort}}
}
\value{
{A sorted matrix}
}
\description{
This function can use the sort.list function in R. The reason
for using it is that one wants the sort to carry along all columns.
}
\examples{

set.seed(30)
x=matrix(sample(1:50),ncol=5)
y=sort_matrix(x,3);y
}
