% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmze.R
\name{symmze}
\alias{symmze}
\title{Replace asymmetric matrix by max of abs values of [i,j] or [j,i] elements.}
\usage{
symmze(mtx)
}
\arguments{
\item{mtx}{{non-symmetric matrix}}
}
\value{
\item{mtx2}{replace [i,j] and [j,i] by the max of absolute values
with common sign}
}
\description{
It is useful in symmetrizing the gmcmtx0 matrix containing a non-symmetric
generalized correlation matrix.
}
\examples{

\dontrun{
example 
mtx=matrix(1:16,nrow=4)
symmze(mtx)
}#' 
}
\author{
Prof. H. D. Vinod, Economics Dept., Fordham University, NY.
}
